+++ src/main/kotlin/com/loadingbyte/cinecred/project/Project.kt
@@ -148,7 +148,8 @@ data class ContentStyle(
     val headMatchWidth: MatchExtent,
     val headMatchWidthAcrossStyles: PersistentList<String>,
     val headHJustify: HJustify,
-    val headVJustify: HeadTailVJustify,
+    val headVShelve: AppendageVShelve,
+    val headVJustify: AppendageVJustify,
     val headGapPx: Double,
     val hasTail: Boolean,
     val tailLetterStyleName: String,
@@ -156,7 +157,8 @@ data class ContentStyle(
     val tailMatchWidth: MatchExtent,
     val tailMatchWidthAcrossStyles: PersistentList<String>,
     val tailHJustify: HJustify,
-    val tailVJustify: HeadTailVJustify,
+    val tailVShelve: AppendageVShelve,
+    val tailVJustify: AppendageVJustify,
     val tailGapPx: Double
 ) : ListedStyle
 
@@ -176,13 +178,8 @@ enum class BodyLayout { GRID, FLOW, PARAGRAPHS }
 enum class HJustify { LEFT, CENTER, RIGHT }
 enum class VJustify { TOP, MIDDLE, BOTTOM }
 enum class LineHJustify { LEFT, CENTER, RIGHT, FULL_LAST_LEFT, FULL_LAST_CENTER, FULL_LAST_RIGHT, FULL_LAST_FULL }
-
-enum class HeadTailVJustify {
-    FIRST_TOP, FIRST_MIDDLE, FIRST_BOTTOM,
-    OVERALL_MIDDLE,
-    LAST_TOP, LAST_MIDDLE, LAST_BOTTOM
-}
-
+enum class AppendageVShelve { FIRST, OVERALL_MIDDLE, LAST }
+enum class AppendageVJustify { TOP, MIDDLE, BOTTOM, BASELINE }
 enum class MatchExtent { OFF, WITHIN_BLOCK, ACROSS_BLOCKS }
 enum class GridFillingOrder { L2R_T2B, R2L_T2B, T2B_L2R, T2B_R2L }
 enum class GridStructure { FREE, EQUAL_WIDTH_COLS, SQUARE_CELLS }